+++ tests/script/test_sig_hash_segwitv0.py
@@ -15,8 +15,7 @@ test vector at https://github.com/bitcoin/bips/blob/master/bip-0143.mediawiki
 
 from btclib.script import sig_hash
 from btclib.script.witness import Witness
-from btclib.tx.tx import Tx
-from btclib.tx.tx_out import TxOut
+from btclib.tx import Tx, TxOut
 
 
 def test_native_p2wpkh() -> None: