+++ tests/tx/test_out_point.py
@@ -16,8 +16,8 @@ from os import path
 import pytest
 
 from btclib.exceptions import BTClibValueError
+from btclib.tx import Tx
 from btclib.tx.out_point import OutPoint
-from btclib.tx.tx import Tx
 
 
 def test_out_point() -> None: