+++ tests/tx/test_tx.py
@@ -17,9 +17,7 @@ import pytest
 
 from btclib.exceptions import BTClibValueError
 from btclib.script.witness import Witness
-from btclib.tx.tx import Tx
-from btclib.tx.tx_in import OutPoint, TxIn
-from btclib.tx.tx_out import TxOut
+from btclib.tx import OutPoint, Tx, TxIn, TxOut
 
 
 def test_tx() -> None: